backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
google-generativeai==0.3.2
azure-search-documents==11.4.0
azure-core==1.29.5
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
